% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pdb.R
\name{download_file}
\alias{download_file}
\title{Download a file from a given URL to a specified destination.}
\usage{
download_file(url, destination, quiet = FALSE)
}
\arguments{
\item{url}{The URL of the file to be downloaded.}

\item{destination}{The local path where the file will be saved.}

\item{quiet}{If TRUE, suppress download progress messages (default is FALSE).}
}
\value{
Returns 0 if the download is successful, and 1 if an error occurs.
}
\description{
This function downloads a file from a provided URL and saves it to the specified destination.
}
