% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pdb.R
\name{get_pdb}
\alias{get_pdb}
\title{Download Protein Data Bank (PDB) files for given PDB IDs.}
\usage{
get_pdb(pdb_id, path = ".", verbose = TRUE, overwrite = FALSE)
}
\arguments{
\item{pdb_id}{A character vector containing PDB IDs.}

\item{path}{The local directory where PDB files will be saved (default is current working directory).}

\item{verbose}{If TRUE, display download progress messages (default is TRUE).}

\item{overwrite}{If TRUE, overwrite existing files with the same name (default is FALSE).}
}
\value{
Returns a character vector with the paths to the downloaded PDB files.
}
\description{
This function downloads PDB files from the RCSB PDB website for the specified PDB IDs.
}
