% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{Trim leading and trailing white spaces from lines in a character vector}
\usage{
trim(lines)
}
\arguments{
\item{lines}{A character vector containing the lines to be processed.}
}
\value{
A character vector with lines trimmed of leading and trailing whitespaces
and empty lines removed.
}
\description{
This function takes a character vector as input, removes leading and trailing
whitespaces, and drops empty lines.
}
\examples{
\dontrun{
input_lines <- c("test1", " test2 ", "", "test3", " test4 ")
result <- trim(input_lines)
print(result)
}

}
